refactor(main): api.ts をドメイン別ルーターに分割し win 解決を middleware 化
571 行の単一ファイルだった api.ts を src/main/api/ のドメイン別
ルーター(app / settings / packages / core / migration / nicommons /
modList)に分割する。tRPC の配線は trpc.ts に集約し、19 箇所で
繰り返していた BrowserWindow.fromWebContents + not found チェックの
定型を winProcedure middleware に置き換える。あわせて getConfig() の
都度呼び出しをやめ、createContext で config を context に載せる。

importer(renderer/trpc.ts の AppRouter 型と windows.ts)は
ディレクトリ import が index.ts に解決されるため変更なし。挙動同一。
