+++ src/main/api/modList.ts
@@ -1,0 +1,8 @@
+import { updateInfo } from '../services/modList';
+import { t, winProcedure } from './trpc';
+
+export const modListRouter = t.router({
+  updateInfo: winProcedure.mutation(async ({ ctx }) => {
+    await updateInfo(ctx.win, ctx.config);
+  }),
+});
